requirements.txt
Flask==2.1.0
Werkzeug==2.1.2
joblib==1.0.0
pandas==1.4.2
numpy==1.19.5
matplotlib==3.3.4
scikit-learn==0.24.1
tensorflow==2.4.1
Pillow==9.5.0
ssl

